style: spawn to improve readability
