+++ tools/simulator/Cargo.toml
@@ -6,6 +6,7 @@ authors = ["Devdutt Shenoi <devdutt@bytebeam.io>"]
 
 [dependencies]
 fake = { version = "2.5.0", features = ["derive"] }
+flume = "0.10"
 futures-util = "0.3"
 log = "0.4"
 rand = "0.8"
@@ -16,5 +17,5 @@ structopt = "0.3"
 thiserror = "1"
 tokio = { version = "1", features = ["full"] }
 tokio-stream = "0.1"
-tokio-util = { version = "0.7", features = ["codec", "time"] }
+tokio-util = { version = "0.7", features = ["codec"] }
 uplink = { path = "../../uplink" }
